+++ lib/src/dialogs/x_dialog.dart
@@ -12,10 +12,10 @@ import "../../x_containers.dart";
 class XDialog {
   // VARIABLES =================================================================
 
-  /// The text to display at the top of the dialog.
+  /// {@macro x_containers.docs.title}
   final Widget title;
 
-  /// The description of the effect of the dialog.
+  /// {@macro x_containers.docs.content}
   final Widget? content;
 
   /// The text displayed on the cancel button.
@@ -28,7 +28,7 @@ class XDialog {
   /// If it is [null] then the button isn't displayed.
   final String? validateText;
 
-  /// The color of the background of the box.
+  /// {@macro x_containers.docs.color}
   final Color? color;
 
   /// The behavior to execute when the "cancel" button is pressed.
@@ -110,13 +110,15 @@ class XDialog {
           shape:
               RoundedRectangleBorder(borderRadius: xTheme.dialogBorderRadius),
           title: DefaultTextStyle(
-              style:
-                  Theme.of(context).textTheme.titleSmall ?? const TextStyle(),
-              child: title),
+            style: const TextStyle()
+                .merge(Theme.of(context).textTheme.titleMedium),
+            child: title,
+          ),
           content: DefaultTextStyle(
-              style:
-                  Theme.of(context).textTheme.bodyMedium ?? const TextStyle(),
-              child: content ?? const SizedBox()),
+            style:
+                const TextStyle().merge(Theme.of(context).textTheme.bodyMedium),
+            child: content ?? const SizedBox(),
+          ),
           actions: <Widget>[
             Visibility(
               visible: cancelText != null,